Add logging code to replicate and pmap
This commit adds optional hooks to pmap. The first executes code
after a worker is assigned to the task but before the task is
executed on that process. The second hook executes after the task
terminates and is returned to the main process.

These hooks are used to make a version of replicate that writes
to logfiles before and after each process is executed.