Add code to save state of RNG
This commit adds an optional random number generator as an argument to
replicate. If it is provided, the state of the RNG will be saved before
each call of 'dgp' and will be saved at the end of the simulations.
This should make debugging easier and also make it possible to extend
the simulations from their final stopping point.